requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
python-telegram-bot==22.0
transformers
peft
nest_asyncio
beautifulsoup4
duckduckgo-search
datasets
pydantic
accelerate
scikit-learn
sqlite-utils
requests
pymupdf
python-docx
gradio
numpy<2.0
faiss-cpu
sentence-transformers
woocommerce
rapidfuzz>=3.0.0
python-telegram-bot
